Fix bad texture coordinates in raster-12
The Cube Model defines the triangle vertexes forming its top face in a
different way to its other faces. It makes no difference for the demos
where the cube faces are solid colors, but it is a problem in
raster-12.html, when textures are introduced. It results in one of the
top face triangles displaying the wrong part of the texture.

Fix the problem by reordering the triangle vertexes on the top cube
face. Although the problem is only visible on raster-12.html, make
similar changes to raster-06.html to raster-12.html so the definition
of cube vertexes remains consistent.